Correctly set the path to the include directory (#7)
Extract the include directory by taking the shortest parent directory
of all the files in the included filegroup instead of assuming the first
one in the filegroup is at the top of the directory.

This assumption does not hold on OS X and is not guaranteed anyway. It
resulted in a wrong include path on Darwin that start failing with
sandboxing because the D toolchain was shipping config file that were
correctly setting those flag but those config file are not present
in the sandbox.