d_test_library: yet another rule for D library
When we build a library for a test, we need to set `version(unittest)`,
since ocasionally the code has some hacks for tests. We can't set this
version directly, so need to pass `-unittest` flag.

This has an unfortunate side effect that the tests inside the library
are also included.