d: support for generated sources
Currently only for d_library and d_source_library.

When we generated sources, we need to patch imports flags, so the
generated files are found by the compiler. In theory, string imports
also need this. But currently I don't generate files I want to import as
strings (though I will).

Currently there is an issue with `d_library`: it doesn't accumulate
transitive imports, so we if A imports from B and B imports from C
and C requires non-standard import location, this won't work.

This works with d_source_library though.