d_library/d_source_library: support transitive imports
That's not entrirely clear if it's good or bad, but let's support
accumulating imports/string imports for all transitive deps.

In most languages, if you need source-level access to a library, it must
be listed in deps explicitly. But in D you can use a template in libA
(which is your direct dep), which in turn uses a template from libB
(which is a direct dep of libA). So you should either add libB to be a
direct dependency, or rely on the build system to collect transitive
sources and imports. Before this change we do collect sources, but not
imports, this is a weird situation half-way through. So let's also
collect transitive imports.

It also simplifies the code a bit.

On the bad side, this breaks the invariant that you can only use APIs
from your direct dependencies: now you can also use APIs from indirect
deps. Which is bad, but that was also true before this change: since we
were collecting transitive sources anyway, we could already use indirect
deps, but only the ones that don't require `-I` flag.