add support for handling generated source files
D relies on paths to find modules. So, if you generate a
source file, it will be under the output directory and D compiler won't
automatically find it. This is solved currently using the `is_generated`
attribute, that forces adding the `-Ioutdir` flag to the compiler.

This works, but not in all cases. Problems start if you need to support
the same module name for both unprocessed and instrumented source.
And then add an extra layer of generation, like generating the
header files. You can have both `liba/moda.d` and `outdir/liba/moda.d`,
but generated header would have to be `outdir/liba/moda.di` in both
cases. And that's not allowed by Bazel.

So, the new `generated_srcs` attribute is added to `d_library` and
`d_binary` rules, allowing to map the generated source files to the
expected source location. We collect this mapping information from all
the dependencies and wrap the compiler invocation with a script that
creates symlinks from the expected location to the actual location.
This way, we can generate files with modified names, and still have
the compiler find them.

Passing that dictionary is a bit cumbersome, but the idea is it will be
mostly used by the bazel macros and not by the users directly.

We can probably remove the `is_generated` attribute now.