d_lib: add hdrs and exports
The idea is to limit the transitive dependencies: instead of making
all srcs visibile, all expose headers + exports.

Why we differentiate? The plan is to use header generation for exports,
while headers are hand-written.

Headers and exports should be closed under imports: if an export A
imports module B, B **must** be in exports as well, otherwise
compilation will be broken.

For now, exports are kinda useless: unless the library interdependency
graph is not connected, because of that closure requirement, the whole
library will end up in exports. This is going to improve with header
generation, since it can truncate some imports.

If neither hdrs nor exports is set, the whole srcs are exported.

As a results, d_source_library is the same as making d_library only with
headers.