Use llc as a bitcode compiler
This is questionable: most resources say llc is only for debugging
and in production people should use clang as a driver. But clang
has different backend options wrt to ldc, while llc has exactly the
same. In the long run, I guess the right way is to use _ldc_ as a
driver.

But right now it doesn't seem to support .bc -> .o compilation.