feature(link): link with cc_common.link
This is the main change. Instead of linking with D compiler, call
`cc_common.link` for better portability.

This requires some changes, mainly we don't collect libraries anymore,
instead we build `LinkingContext` for rules_cc.

This simplifies handling of C dependencies, but add a bit of wrapping
around D libraries.