feat(compile): check compiler errors and output useful info
The idea is to use iBazel output runner feature, but iBazel only
allows matching on a single line.

Usually this should be enough: seeing something like
```
my/package/file.d: Expected 'x/y/z.d' or 'x/y/z/package.d' ...
```

we can figure out that d_library containing `x/y/z.d` of
`x/y/z/package.d` needs to be added to a target containing
`my/package/file.d`.

However, this is one complication with the cycling dependencies and
their breaker libraries:

Imagine `x/y/z.d` belongs to the lib X, and `my/package/file.d` belongs
to another lib Y. We surely want to add X to the deps of Y.

But that might not be enough. If this is some other lib Z that depends
on Y-cycle-breaker, it _won't_ get Y from X automatically. So we need to
add X or X-cycle-breaker to Z directly. To support this case, we
need to know which target we are compiling currently, and that's exactly
what's wrapper script will gives us with this commit.

If iBazel supports passing the failing target to the output runner, that
would be not necessary.